requirements.txt
# Core Deep Learning
torch>=2.2.0
torchvision>=0.17.0

# Anomaly Detection
anomalib>=1.1.0

# Computer Vision
opencv-python>=4.9.0
Pillow>=10.2.0

# Data & Metrics
numpy>=1.26.0
scikit-learn>=1.4.0

# Config & Utils
pyyaml>=6.0.1

# ONNX Export
onnx>=1.16.0
onnxruntime-gpu>=1.17.0

# Logging & Progress
tqdm>=4.66.0


## Run this BEFORE pip install -r requirements.txt, if you have GPU.
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121

# pip install -r requirements.txt 

# Quick verify after installation:
# python -c "import torch; print('GPU Available:', torch.cuda.is_available())"

